# src/api/Bee2Bee-1.0.0-swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Algorithm-based Employee Exchange Platform Bee2Bee API",
    "version" : "1.0.0",
    "title" : "Algorithm-based Employee Exchange Platform Bee2Bee API",
    "contact" : {
      "url" : "https://github.com/BeeToBeeOrg/Bee2Bee_backend"
    },
    "license" : {
      "name" : "MIT"
    }
  },
...
  "paths" : {
    "/user" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Create a new user",
        "description" : "",
        "operationId" : "addUser",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "newUser",
          "description" : "New User Infos",
          "schema" : {
            "type" : "object",
            "required" : [ "email", "company_id", "name", "secret" ],
            "properties" : {
              "email" : {
                "type" : "string"
              },
              "secret" : {
                "type" : "string"
              },
              "name" : {
                "type" : "string"
              },
              "company_id" : {
                "type" : "integer"
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "return the newly created user object",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          },
          "400" : {
            "description" : "bad input parameter"
          },
          "405" : {
            "description" : "Validation exception"
          }
        }
      }
    },
    "/user/auth" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Auth a user",
        "description" : "",
        "operationId" : "authUser",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "authUser",
          "description" : "Auth the user credentials",
          "schema" : {
            "type" : "object",
            "required" : [ "email", "secret" ],
            "properties" : {
              "email" : {
                "type" : "string"
              },
              "secret" : {
                "type" : "string"
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "login success",
            "schema" : {
              "type" : "object",
              "properties" : {
                "token" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          },
          "403" : {
            "description" : "login failed"
          },
          "405" : {
            "description" : "Validation exception"
          }
        }
      }
    },
    "/user/{token}" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "Find user by token",
        "description" : "Returns a single user",
        "operationId" : "getUserByToken",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "token",
          "in" : "path",
          "description" : "Session Token of User to return",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          },
          "400" : {
            "description" : "Invalid Token supplied"
          },
          "404" : {
            "description" : "User not found"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "bee2bee_auth" : {
      "type" : "oauth2",
      "authorizationUrl" : "https://.../oauth/dialog",
      "flow" : "implicit"
    },
    "api_key" : {
      "type" : "apiKey",
      "name" : "api_key",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "User" : {
      "required" : [ "email", "id", "company_id" ],
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "company_id" : {
          "type" : "integer"
        },
        "name" : {
          "type" : "string",
          "example" : "John Doe"
        },
        "email" : {
          "type" : "string",
          "example" : "test@test.de"
        }
      }
    },
...
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/Bee2Bee/Bee2Bee/1.0.0",
  "schemes" : [ "https" ]
}